add RBAC feature

+++ backend/controllers/CommentController.php
@@ -3,11 +3,12 @@
 namespace backend\controllers;
 
 use Yii;
-use common\models\Comment;
-use common\models\CommentSearch;
 use yii\web\Controller;
-use yii\web\NotFoundHttpException;
+use common\models\Comment;
 use yii\filters\VerbFilter;
+use common\models\CommentSearch;
+use yii\web\NotFoundHttpException;
+use yii\web\ForbiddenHttpException;
 
 /**
  * CommentController implements the CRUD actions for Comment model.
@@ -65,12 +66,18 @@ class CommentController extends Controller
     /**
      * Updates an existing Comment model.
      * If update is successful, the browser will be redirected to the 'view' page.
+     *
      * @param integer $id
      * @return mixed
+     * @throws ForbiddenHttpException
      * @throws NotFoundHttpException if the model cannot be found
      */
     public function actionUpdate($id)
     {
+        if (!Yii::$app->user->can('updateComment')) {
+            throw new ForbiddenHttpException('暂无权限执行此操作');
+        }
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -85,12 +92,20 @@ class CommentController extends Controller
     /**
      * Deletes an existing Comment model.
      * If deletion is successful, the browser will be redirected to the 'index' page.
+     *
      * @param integer $id
      * @return mixed
+     * @throws ForbiddenHttpException
      * @throws NotFoundHttpException if the model cannot be found
+     * @throws \Throwable
+     * @throws \yii\db\StaleObjectException
      */
     public function actionDelete($id)
     {
+        if (!Yii::$app->user->can('deleteComment')) {
+            throw new ForbiddenHttpException('暂无权限执行此操作');
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
